+++ src/server/game/Globals/ObjectMgr.h
@@ -1376,9 +1376,9 @@ public:
 
     void LoadInstanceSavedGameobjectStateData();
     bool FindInstanceSavedGameobjectState(uint32 id, uint32 guid);
-    unsigned short GetInstanceSavedGameobjectState(uint32 id, uint32 guid);
-    void SetInstanceSavedGameobjectState(uint32 id, uint32 guid, unsigned short state);
-    void NewInstanceSavedGameobjectState(uint32 id, uint32 guid, unsigned short state);
+    uint8 GetInstanceSavedGameobjectState(uint32 id, uint32 guid);
+    void SetInstanceSavedGameobjectState(uint32 id, uint32 guid, uint8 state);
+    void NewInstanceSavedGameobjectState(uint32 id, uint32 guid, uint8 state);
 private:
     // first free id for selected id type
     uint32 _auctionId; // pussywizard: accessed by a single thread
